src/requirements_torch.txt
absl-py==0.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.1
catboost==1.0.3
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cloudpickle==2.0.0
cryptography==2.1.4
cycler==0.11.0
dask==2021.3.0
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
faiss-cpu==1.7.1.post3
flatbuffers==1.12
fsspec==2021.11.1
future==0.18.2
gast==0.3.3
gensim==4.1.2
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
lightgbm==3.3.1
llvmlite==0.36.0
locket==0.2.1
lxml==4.7.1
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mock==4.0.3
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.5.0
portpicker==1.3.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
Pygments==2.10.0
PyGObject==3.26.1
pyparsing==3.0.6
pyrsistent==0.18.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytrec-eval==0.5
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==1.11.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
smart-open==5.2.1
swifter==1.0.9
tenacity==8.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-serving-api-gpu @ file:///tmp/pip/tensorflow_serving_api_gpu-2.4.1-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
xgboost==1.5.2
zipp==3.4.0
